add dev e2e tests for API routes
